revert workspace jest test changes
